dns/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
dnspython==2.6.1
    # via -r requirements.in
tomlkit==0.12.5
    # via -r requirements.in
